requirements.txt
pyweatherbit
python_weather
openpyxl